+++ src/index.ts
@@ -5,9 +5,12 @@ export * from "./soroban_rpc";
 export { ContractSpec } from "./contract_spec";
 
 // stellar-sdk classes to expose
-export { Server } from "./server";
+export { Server, Durability } from "./server";
 export { AxiosClient, version } from "./axios";
 export * from "./transaction";
+
+// export is necessary for testing, but it may be useful for others
+export { parseRawSimulation } from "./parsers";
 
 // expose classes and functions from stellar-base
 export * from "stellar-base";