+++ test/unit/server/get_account_test.js
@@ -1,8 +1,6 @@
-const MockAdapter = require('axios-mock-adapter');
+const { Account, Keypair, StrKey, xdr } = SorobanClient;
 
 describe('Server#getAccount', function () {
-  const { Account, StrKey, xdr } = SorobanClient;
-
   beforeEach(function () {
     this.server = new SorobanClient.Server(serverUrl);
     this.axiosMock = sinon.mock(AxiosClient);
@@ -13,35 +11,30 @@ describe('Server#getAccount', function () {
     this.axiosMock.restore();
   });
 
-  it('requests the correct method', function (done) {
-    const address = 'GBZXN7PIRZGNMHGA7MUUUF4GWPY5AYPV6LY4UV2GL6VJGIQRXFDNMADI';
-    const accountId = xdr.PublicKey.publicKeyTypeEd25519(
-      StrKey.decodeEd25519PublicKey(address)
-    );
+  const address = 'GBZXN7PIRZGNMHGA7MUUUF4GWPY5AYPV6LY4UV2GL6VJGIQRXFDNMADI';
+  const accountId = Keypair.fromPublicKey(address).xdrPublicKey();
+  const key = xdr.LedgerKey.account(new xdr.LedgerKeyAccount({ accountId }));
+  const accountEntry =
+    'AAAAAAAAAABzdv3ojkzWHMD7KUoXhrPx0GH18vHKV0ZfqpMiEblG1g3gtpoE608YAAAAAAAAAAEAAAAAAAAAAAAAAAAAAAAAAQAAAAAAAAAAAAABAAAAAAAAAAAAAAAAAAAAAAAAAAIAAAAAAAAAAAAAAAAAAAADAAAAAAAAAAQAAAAAY9D8iA';
 
+  it('requests the correct method', function (done) {
     this.axiosMock
       .expects('post')
       .withArgs(serverUrl, {
         jsonrpc: '2.0',
         id: 1,
         method: 'getLedgerEntries',
-        params: [
-          [
-            xdr.LedgerKey.account(
-              new xdr.LedgerKeyAccount({
-                accountId
-              })
-            ).toXDR('base64')
-          ]
-        ]
+        params: [[key.toXDR('base64')]]
       })
       .returns(
         Promise.resolve({
           data: {
             result: {
+              latestLedger: 0,
               entries: [
                 {
-                  xdr: 'AAAAAAAAAABzdv3ojkzWHMD7KUoXhrPx0GH18vHKV0ZfqpMiEblG1g3gtpoE608YAAAAAAAAAAEAAAAAAAAAAAAAAAAAAAAAAQAAAAAAAAAAAAABAAAAAAAAAAAAAAAAAAAAAAAAAAIAAAAAAAAAAAAAAAAAAAADAAAAAAAAAAQAAAAAY9D8iA'
+                  key: key.toXDR('base64'),
+                  xdr: accountEntry
                 }
               ]
             }
@@ -71,20 +64,13 @@ describe('Server#getAccount', function () {
         jsonrpc: '2.0',
         id: 1,
         method: 'getLedgerEntries',
-        params: [
-          [
-            xdr.LedgerKey.account(
-              new xdr.LedgerKeyAccount({
-                accountId
-              })
-            ).toXDR('base64')
-          ]
-        ]
+        params: [[key.toXDR('base64')]]
       })
       .returns(
         Promise.resolve({
           data: {
             result: {
+              latestLedger: 0,
               entries: null
             }
           }